Enhance file size checking script and diagnostics
- Update `check_file_sizes.sh` to include an option for excluding test files by default, improving usability.
- Introduce a percentage calculation for "good" files, providing an overall rating based on file sizes, enhancing feedback for developers.
- Modify diagnostic messages to reflect the new rating system, offering clearer guidance on file size evaluations.
- Add new helper functions in the `source` package for span manipulation, improving code clarity and maintainability.

This commit improves the file size checking process, making it more informative and user-friendly, while also enhancing the underlying code structure.
