+++ internal/source/span.go
@@ -35,6 +35,7 @@ func (s Span) Cover(other Span) Span {
 	return s
 }
 
+// ShiftLeft сдвигает span налево на n символов
 func (s Span) ShiftLeft(n uint32) Span {
 	if n > s.Start {
 		return s
@@ -46,6 +47,7 @@ func (s Span) ShiftLeft(n uint32) Span {
 	}
 }
 
+// ShiftRight сдвигает span направо на n символов
 func (s Span) ShiftRight(n uint32) Span {
 	if n > s.End - s.Start {
 		return s
@@ -54,5 +56,25 @@ func (s Span) ShiftRight(n uint32) Span {
 		File:  s.File,
 		Start: s.Start + n,
 		End:   s.End + n,
+	}
+}
+
+// ZeroideToStart возвращает span, где start == end == изначальный start
+// используется для Insert операций
+func (s Span) ZeroideToStart() Span {
+	return Span{
+		File:  s.File,
+		Start: s.Start,
+		End:   s.Start,
+	}
+}
+
+// ZeroideToEnd возвращает span, где start == end == изначальный end
+// используется для Insert операций
+func (s Span) ZeroideToEnd() Span {
+	return Span{
+		File:  s.File,
+		Start: s.End,
+		End:   s.End,
 	}
 }
