+++ testdata/test_fixes/let_fixes/let_arr.sg
@@ -1,0 +1,2 @@
+let a: int[]; // нет ошибки
+let b: int[; //
